Refactored AvStream. Introduced classes Node, StreamInfo, StreamQueue, Decoder.
Complete refactoring of AvStream with a new design of the stream flow and threads.
Now everything is a Node and a Node has as many input queues as it has input streams.
Looks more familiar (resembles gstreamer and NMM somehow), but with some special design
features: you can select each stream of two nodes to be connected. And this can be done
dynamically at runtime (having the still-to-come big TS stream remuxer in the DVB server in mind).
It compiles but segfaults, what did ya expect ...?