Frames got numbers and queues got names. Improved logging.
Frame number and pts are stored as int64_t and display a nice name with number, stream name and pts shown.
Queues are named, to better distinguish streams, queues, and threads in the log. Logging output has really improved.
This depicts wonderfully the still decoupled threads in the video sink, resulting in the timed display of the same
frame in sequence.