ADD ByteQueue and SdlAudioSink.
ByteQueue is a blocking queue with buffered read and write. It is used in AudioSink
to queue the audio sample stream to the PCM buffer. Currently needs to catch a
Poco::SystemException when semaphore is already at max and a thread tries to set
the semaphore, again. Discovered, that I also need to set -malign-double, when
compiling code against libommavstream.