ADD initial implementation of DirectoryContainer while further refactoring MediaObject.
DirectoryContainer is the second backend for the new (Abstract)MediaContainer.
It uses a 'torch' strategy to browse a directory on a filesystem, which means
it doesn't allocate meta data in memory or a cache or a database, but generates
meta data on the fly. This is the 'more cpu, less memory' strategy in the
trade off between cpu time and memory usage. This strategy fits well for
directories with many media files on small media servers. Without support
for search functionality, this completely eliminates any filesystem scanning
and is able to deliver any media file on the filesystem right from the start
of the media server.