CHG removed device container code from device dispatcher / implementation.
On server side, device container is now created on application level, where also devices
are created and added to the device container. Also, not the devices are started
but the device container, because it owns the http and ssdp sockets. Later, code should
be added to properly register / unregister devices in the device container, sending out
corresponding ssdp messages.