FIX devices appeared twice in device manager.
DeviceManager has a container of DeviceContainers. When a device container
pops up through SSDP, it is added to this container. SSDP messages may appear
double. Device containers are added to the container with uuid of their
root device as key, so they are unique. Formerly, only root devices were
recogniced by the controller, no subdevices.