CHG simplified creating of controller devices after discovery.
Now, a specialized device (e.g. MediaRenderer) is created without
setting any pointers or adding controller code. Then the spezialized
device is inserted into the device tree, replacing the general device
that was discovered. Then the specialized device code is created,
initialized and attached to the specialized device. Finally, the
new device is added to the device group and eventing is activated.