CHG replaced EventLoop and MainWindow with Gui::Application.
UIKit builds up the view hierarchy after entering the event loop.
This is accomplished with the callback applicationDidFinishLaunching.
As a consequence we can't allocate any views before starting the main
loop, so we need a factory method for creating the views and main window
and event loop in one class.