FIX setting the volume also triggered an event that in turn set the volume.
Slider is a somewhat special control because it controls and displays its
value with the same view element. Therefore a feedback loop can occure when
changing the value which in turn updates the value from an external thread.
Introduced a passive mode (value is set external) and active mode (value is
set by gui main thread) for the slider.