ADD config option to check for object modifications or not.
In most cases, checking for new and removed object paths is sufficient,
as access of an object resource is not affected by the rest of the object's
meta data. If modifications are checked, then changing the title of an
object, for example, triggers a change of the system update id. Second
mode is more expensive when polling the system update id and updating
the object cache.