FIX workaround for msearch packet lost when iphone wifi is reachable but network not.
Introduced a delay of 1000ms after iphone tells that wifi is reachable. Without that
the msearch udp packet is lost. This should be replaced by a safer method, as it may
take longer until the connection to the router is established. Reachability framework
only tells availability of connection ... the Darwin backend uses another framework.